+++ RELEASE_NOTES.md
@@ -1,2 +1,4 @@
 v0.97.3
 
+- Add AuthUsernamesService.deleteUsernameHash()
+- Add AuthUsernamesService.deleteUsernameLink()